Bump versions and convert unit tests
