+++ src/tests/controllers/api/damageTypeController.test.ts
@@ -1,10 +1,11 @@
+import mockingoose from 'mockingoose';
 import { mockNext, mockRequest, mockResponse } from '../../support/requestHelpers.js';
 
-import Equipment from '../../../models/equipment/index.js';
-import EquipmentController from '../../../controllers/api/equipmentController.js';
-import mockingoose from 'mockingoose';
+import DamageType from '../../../models/damageType/index.js';
+import DamageTypeController from '../../../controllers/api/damageTypeController.js';
+import { MockResponse } from '../../support/types.d';
 
-let response;
+let response: MockResponse;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
@@ -13,27 +14,27 @@ beforeEach(() => {
 describe('index', () => {
   const findDoc = [
     {
-      index: 'abacus',
-      name: 'Abacus',
-      url: '/api/equipment/abacus',
+      index: 'acid',
+      name: 'Acid',
+      url: '/api/damage-types/acid',
     },
     {
-      index: 'acid-vial',
-      name: 'Acid (vial)',
-      url: '/api/equipment/acid-vial',
+      index: 'bludgeoning',
+      name: 'Bludgeoning',
+      url: '/api/damage-types/bludgeoning',
     },
     {
-      index: 'alchemists-fire-flask',
-      name: "Alchemist's fire (flask)",
-      url: '/api/equipment/alchemists-fire-flask',
+      index: 'cold',
+      name: 'Cold',
+      url: '/api/damage-types/cold',
     },
   ];
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(Equipment).toReturn(findDoc, 'find');
+    mockingoose(DamageType).toReturn(findDoc, 'find');
 
-    await EquipmentController.index(request, response, mockNext);
+    await DamageTypeController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -41,9 +42,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Equipment).toReturn(error, 'find');
+      mockingoose(DamageType).toReturn(error, 'find');
 
-      await EquipmentController.index(request, response, mockNext);
+      await DamageTypeController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -54,18 +55,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'abacus',
-    name: 'Abacus',
-    url: '/api/equipment/abacus',
+    index: 'acid',
+    name: 'Acid',
+    url: '/api/damage-types/acid',
   };
 
-  const showParams = { index: 'abacus' };
+  const showParams = { index: 'acid' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(Equipment).toReturn(findOneDoc, 'findOne');
+    mockingoose(DamageType).toReturn(findOneDoc, 'findOne');
 
-    await EquipmentController.show(request, response, mockNext);
+    await DamageTypeController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -73,11 +74,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(Equipment).toReturn(null, 'findOne');
+      mockingoose(DamageType).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await EquipmentController.show(invalidRequest, response, mockNext);
+      await DamageTypeController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -88,9 +89,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Equipment).toReturn(error, 'findOne');
+      mockingoose(DamageType).toReturn(error, 'findOne');
 
-      await EquipmentController.show(request, response, mockNext);
+      await DamageTypeController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
