+++ src/tests/controllers/api/equipmentCategoryController.test.ts
@@ -1,10 +1,11 @@
+import mockingoose from 'mockingoose';
 import { mockNext, mockRequest, mockResponse } from '../../support/requestHelpers.js';
 
-import Trait from '../../../models/trait/index.js';
-import TraitController from '../../../controllers/api/traitController.js';
-import mockingoose from 'mockingoose';
+import EquipmentCategory from '../../../models/equipmentCategory/index.js';
+import EquipmentCategoryController from '../../../controllers/api/equipmentCategoryController.js';
+import { MockResponse } from '../../support/types.d';
 
-let response;
+let response: MockResponse;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
@@ -13,27 +14,27 @@ beforeEach(() => {
 describe('index', () => {
   const findDoc = [
     {
-      index: 'artificers-lore',
-      name: "Artificer's Lore",
-      url: '/api/traits/artificers-lore',
+      index: 'adventuring-gear',
+      name: 'Adventuring Gear',
+      url: '/api/equipment-categories/adventuring-gear',
     },
     {
-      index: 'brave',
-      name: 'Brave',
-      url: '/api/traits/brave',
+      index: 'armor',
+      name: 'Armor',
+      url: '/api/equipment-categories/armor',
     },
     {
-      index: 'breath-weapon',
-      name: 'Breath Weapon',
-      url: '/api/traits/breath-weapon',
+      index: 'artisans-tools',
+      name: "Artisan's Tools",
+      url: '/api/equipment-categories/artisans-tools',
     },
   ];
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(Trait).toReturn(findDoc, 'find');
+    mockingoose(EquipmentCategory).toReturn(findDoc, 'find');
 
-    await TraitController.index(request, response, mockNext);
+    await EquipmentCategoryController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -41,9 +42,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Trait).toReturn(error, 'find');
+      mockingoose(EquipmentCategory).toReturn(error, 'find');
 
-      await TraitController.index(request, response, mockNext);
+      await EquipmentCategoryController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -54,18 +55,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'abyssal',
-    name: 'Abyssal',
-    url: '/api/traits/abyssal',
+    index: 'adventuring-gear',
+    name: 'Adventuring Gear',
+    url: '/api/equipment-categories/adventuring-gear',
   };
 
-  const showParams = { index: 'abyssal' };
+  const showParams = { index: 'adventuring-gear' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(Trait).toReturn(findOneDoc, 'findOne');
+    mockingoose(EquipmentCategory).toReturn(findOneDoc, 'findOne');
 
-    await TraitController.show(request, response, mockNext);
+    await EquipmentCategoryController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -73,11 +74,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(Trait).toReturn(null, 'findOne');
+      mockingoose(EquipmentCategory).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await TraitController.show(invalidRequest, response, mockNext);
+      await EquipmentCategoryController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -88,9 +89,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Trait).toReturn(error, 'findOne');
+      mockingoose(EquipmentCategory).toReturn(error, 'findOne');
 
-      await TraitController.show(request, response, mockNext);
+      await EquipmentCategoryController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
