+++ src/tests/controllers/api/featController.test.ts
@@ -1,10 +1,11 @@
+import mockingoose from 'mockingoose';
 import { mockNext, mockRequest, mockResponse } from '../../support/requestHelpers.js';
 
-import Feature from '../../../models/feature/index.js';
-import FeatureController from '../../../controllers/api/featureController.js';
-import mockingoose from 'mockingoose';
+import Feat from '../../../models/feat/index.js';
+import FeatController from '../../../controllers/api/featController.js';
+import { MockResponse } from '../../support/types.d';
 
-let response;
+let response: MockResponse;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
@@ -13,27 +14,18 @@ beforeEach(() => {
 describe('index', () => {
   const findDoc = [
     {
-      index: 'action-surge-1-use',
-      name: 'Action Surge (1 use)',
-      url: '/api/features/action-surge-1-use',
-    },
-    {
-      index: 'action-surge-2-uses',
-      name: 'Action Surge (2 uses)',
-      url: '/api/features/action-surge-2-uses',
-    },
-    {
-      index: 'additional-magical-secrets',
-      name: 'Additional Magical Secrets',
-      url: '/api/features/additional-magical-secrets',
+      index: 'grappler',
+      name: 'Grappler',
+      url: '/api/feats/grappler',
     },
   ];
+
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(Feature).toReturn(findDoc, 'find');
+    mockingoose(Feat).toReturn(findDoc, 'find');
 
-    await FeatureController.index(request, response, mockNext);
+    await FeatController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -41,9 +33,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Feature).toReturn(error, 'find');
+      mockingoose(Feat).toReturn(error, 'find');
 
-      await FeatureController.index(request, response, mockNext);
+      await FeatController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -54,18 +46,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'action-surge-1-use',
-    name: 'Action Surge (1 use)',
-    url: '/api/features/action-surge-1-use',
+    index: 'grappler',
+    name: 'Grappler',
+    url: '/api/feats/grappler',
   };
 
-  const showParams = { index: 'action-surge-1-use' };
+  const showParams = { index: 'grappler' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(Feature).toReturn(findOneDoc, 'findOne');
+    mockingoose(Feat).toReturn(findOneDoc, 'findOne');
 
-    await FeatureController.show(request, response, mockNext);
+    await FeatController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -73,11 +65,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(Feature).toReturn(null, 'findOne');
+      mockingoose(Feat).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await FeatureController.show(invalidRequest, response, mockNext);
+      await FeatController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -88,9 +80,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Feature).toReturn(error, 'findOne');
+      mockingoose(Feat).toReturn(error, 'findOne');
 
-      await FeatureController.show(request, response, mockNext);
+      await FeatController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
