+++ src/tests/controllers/api/languageController.test.ts
@@ -1,10 +1,11 @@
+import mockingoose from 'mockingoose';
 import { mockNext, mockRequest, mockResponse } from '../../support/requestHelpers.js';
 
-import AbilityScore from '../../../models/abilityScore/index.js';
-import AbilityScoreController from '../../../controllers/api/abilityScoreController.js';
-import mockingoose from 'mockingoose';
+import Language from '../../../models/language/index.js';
+import LanguageController from '../../../controllers/api/languageController.js';
+import { MockResponse } from '../../support/types.d';
 
-let response;
+let response: MockResponse;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
@@ -13,27 +14,27 @@ beforeEach(() => {
 describe('index', () => {
   const findDoc = [
     {
-      index: 'str',
-      name: 'STR',
-      url: '/api/ability-scores/str',
+      index: 'abyssal',
+      name: 'Abyssal',
+      url: '/api/languages/abyssal',
     },
     {
-      index: 'dex',
-      name: 'DEX',
-      url: '/api/ability-scores/dex',
+      index: 'celestial',
+      name: 'Celestial',
+      url: '/api/languages/celestial',
     },
     {
-      index: 'con',
-      name: 'CON',
-      url: '/api/ability-scores/con',
+      index: 'common',
+      name: 'Common',
+      url: '/api/languages/common',
     },
   ];
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(AbilityScore).toReturn(findDoc, 'find');
+    mockingoose(Language).toReturn(findDoc, 'find');
 
-    await AbilityScoreController.index(request, response, mockNext);
+    await LanguageController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -41,9 +42,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(AbilityScore).toReturn(error, 'find');
+      mockingoose(Language).toReturn(error, 'find');
 
-      await AbilityScoreController.index(request, response, mockNext);
+      await LanguageController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -54,18 +55,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'str',
-    name: 'STR',
-    url: '/api/ability-scores/str',
+    index: 'abyssal',
+    name: 'Abyssal',
+    url: '/api/languages/abyssal',
   };
 
-  const showParams = { index: 'str' };
+  const showParams = { index: 'abyssal' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(AbilityScore).toReturn(findOneDoc, 'findOne');
+    mockingoose(Language).toReturn(findOneDoc, 'findOne');
 
-    await AbilityScoreController.show(request, response, mockNext);
+    await LanguageController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -73,11 +74,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(AbilityScore).toReturn(null, 'findOne');
+      mockingoose(Language).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await AbilityScoreController.show(invalidRequest, response, mockNext);
+      await LanguageController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -88,9 +89,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(AbilityScore).toReturn(error, 'findOne');
+      mockingoose(Language).toReturn(error, 'findOne');
 
-      await AbilityScoreController.show(request, response, mockNext);
+      await LanguageController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
