+++ src/tests/controllers/api/proficiencyController.test.ts
@@ -1,11 +1,11 @@
-import * as SpellController from '../../../controllers/api/spellController.js';
-
-import { mockNext, mockRequest, mockResponse } from '../../support/requestHelpers.js';
-import Spell from '../../../models/spell/index.js';
 import mockingoose from 'mockingoose';
-import { jest } from '@jest/globals';
+import { mockNext, mockRequest, mockResponse } from '../../support/requestHelpers.js';
 
-let response;
+import Proficiency from '../../../models/proficiency/index.js';
+import ProficiencyController from '../../../controllers/api/proficiencyController.js';
+import { MockResponse } from '../../support/types.d';
+
+let response: MockResponse;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
@@ -14,31 +14,27 @@ beforeEach(() => {
 describe('index', () => {
   const findDoc = [
     {
-      index: 'acid-arrow',
-      name: 'Acid Arrow',
-      url: '/api/spells/acid-arrow',
+      index: 'alchemists-supplies',
+      name: "Alchemist's supplies",
+      url: '/api/proficiencies/alchemists-supplies',
     },
     {
-      index: 'acid-splash',
-      name: 'Acid Splash',
-      url: '/api/spells/acid-splash',
+      index: 'all-armor',
+      name: 'All armor',
+      url: '/api/proficiencies/all-armor',
     },
     {
-      index: 'aid',
-      name: 'Aid',
-      url: '/api/spells/aid',
+      index: 'bagpipes',
+      name: 'Bagpipes',
+      url: '/api/proficiencies/bagpipes',
     },
   ];
   const request = mockRequest({ query: {} });
 
-  afterEach(() => {
-    jest.restoreAllMocks();
-  });
-
   it('returns a list of objects', async () => {
-    mockingoose(Spell).toReturn(findDoc, 'find');
+    mockingoose(Proficiency).toReturn(findDoc, 'find');
 
-    await SpellController.index(request, response, mockNext);
+    await ProficiencyController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -46,9 +42,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Spell).toReturn(error, 'find');
+      mockingoose(Proficiency).toReturn(error, 'find');
 
-      await SpellController.index(request, response, mockNext);
+      await ProficiencyController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -59,18 +55,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'acid-arrow',
-    name: 'Acid Arrow',
-    url: '/api/spells/acid-arrow',
+    index: 'alchemists-supplies',
+    name: "Alchemist's supplies",
+    url: '/api/proficiencies/alchemists-supplies',
   };
 
-  const showParams = { index: 'acid-arrow' };
+  const showParams = { index: 'alchemists-supplies' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(Spell).toReturn(findOneDoc, 'findOne');
+    mockingoose(Proficiency).toReturn(findOneDoc, 'findOne');
 
-    await SpellController.show(request, response, mockNext);
+    await ProficiencyController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -78,24 +74,24 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(Spell).toReturn(null, 'findOne');
+      mockingoose(Proficiency).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await SpellController.show(invalidRequest, response, mockNext);
+      await ProficiencyController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
-      expect(mockNext).toHaveBeenCalledWith();
+      expect(mockNext).toHaveBeenCalled();
     });
   });
 
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Spell).toReturn(error, 'findOne');
+      mockingoose(Proficiency).toReturn(error, 'findOne');
 
-      await SpellController.show(request, response, mockNext);
+      await ProficiencyController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
