+++ src/tests/controllers/api/raceController.test.ts
@@ -1,13 +1,15 @@
-import * as SubraceController from '../../../controllers/api/subraceController.js';
+import mockingoose from 'mockingoose';
+import * as RaceController from '../../../controllers/api/raceController.js';
 
 import { mockNext, mockRequest, mockResponse } from '../../support/requestHelpers.js';
 
 import Proficiency from '../../../models/proficiency/index.js';
+import Race from '../../../models/race/index.js';
 import Subrace from '../../../models/subrace/index.js';
 import Trait from '../../../models/trait/index.js';
-import mockingoose from 'mockingoose';
+import { MockResponse } from '../../support/types.d';
 
-let response;
+let response: MockResponse;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
@@ -16,32 +18,27 @@ beforeEach(() => {
 describe('index', () => {
   const findDoc = [
     {
-      index: 'high-elf',
-      name: 'High Elf',
-      url: '/api/subraces/high-elf',
+      index: 'dragonborn',
+      name: 'Dragonborn',
+      url: '/api/races/dragonborn',
     },
     {
-      index: 'hill-dwarf',
-      name: 'Hill Dwarf',
-      url: '/api/subraces/hill-dwarf',
+      index: 'dwarf',
+      name: 'Dwarf',
+      url: '/api/races/dwarf',
     },
     {
-      index: 'lightfoot-halfling',
-      name: 'Lightfoot Halfling',
-      url: '/api/subraces/lightfoot-halfling',
-    },
-    {
-      index: 'rock-gnome',
-      name: 'Rock Gnome',
-      url: '/api/subraces/rock-gnome',
+      index: 'elf',
+      name: 'Elf',
+      url: '/api/races/elf',
     },
   ];
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(Subrace).toReturn(findDoc, 'find');
+    mockingoose(Race).toReturn(findDoc, 'find');
 
-    await SubraceController.index(request, response, mockNext);
+    await RaceController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -49,9 +46,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Subrace).toReturn(error, 'find');
+      mockingoose(Race).toReturn(error, 'find');
 
-      await SubraceController.index(request, response, mockNext);
+      await RaceController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -62,18 +59,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'high-elf',
-    name: 'High Elf',
-    url: '/api/subraces/high-elf',
+    index: 'dragonborn',
+    name: 'Dragonborn',
+    url: '/api/races/dragonborn',
   };
 
-  const showParams = { index: 'high-elf' };
+  const showParams = { index: 'dragonborn' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(Subrace).toReturn(findOneDoc, 'findOne');
+    mockingoose(Race).toReturn(findOneDoc, 'findOne');
 
-    await SubraceController.show(request, response, mockNext);
+    await RaceController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -81,11 +78,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(Subrace).toReturn(null, 'findOne');
+      mockingoose(Race).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await SubraceController.show(invalidRequest, response, mockNext);
+      await RaceController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -96,9 +93,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Subrace).toReturn(error, 'findOne');
+      mockingoose(Race).toReturn(error, 'findOne');
 
-      await SubraceController.show(request, response, mockNext);
+      await RaceController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -107,7 +104,40 @@ describe('show', () => {
   });
 });
 
-describe('showTraitsForSubrace', () => {
+describe('showSubracesForRace', () => {
+  const findDoc = [
+    {
+      index: 'high-elf',
+      name: 'High Elf',
+      url: '/api/subraces/high-elf',
+    },
+  ];
+  const showParams = { index: 'dragonborn' };
+  const request = mockRequest({ params: showParams });
+
+  it('returns a list of objects', async () => {
+    mockingoose(Subrace).toReturn(findDoc, 'find');
+
+    await RaceController.showSubracesForRace(request, response, mockNext);
+
+    expect(response.status).toHaveBeenCalledWith(200);
+  });
+
+  describe('when something goes wrong', () => {
+    it('handles the error', async () => {
+      const error = new Error('Something went wrong');
+      mockingoose(Subrace).toReturn(error, 'find');
+
+      await RaceController.showSubracesForRace(request, response, mockNext);
+
+      expect(response.status).not.toHaveBeenCalled();
+      expect(response.json).not.toHaveBeenCalled();
+      expect(mockNext).toHaveBeenCalledWith(error);
+    });
+  });
+});
+
+describe('showTraitsForRace', () => {
   const findDoc = [
     {
       index: 'darkvision',
@@ -120,9 +150,9 @@ describe('showTraitsForSubrace', () => {
       url: '/api/traits/elf-weapon-training',
     },
     {
-      index: 'extra-language',
-      name: 'Extra Language',
-      url: '/api/traits/extra-language',
+      index: 'fey-ancestry',
+      name: 'Fey Ancestry',
+      url: '/api/traits/fey-ancestry',
     },
   ];
   const showParams = { index: 'elf' };
@@ -131,7 +161,7 @@ describe('showTraitsForSubrace', () => {
   it('returns a list of objects', async () => {
     mockingoose(Trait).toReturn(findDoc, 'find');
 
-    await SubraceController.showTraitsForSubrace(request, response, mockNext);
+    await RaceController.showTraitsForRace(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -141,7 +171,7 @@ describe('showTraitsForSubrace', () => {
       const error = new Error('Something went wrong');
       mockingoose(Trait).toReturn(error, 'find');
 
-      await SubraceController.showTraitsForSubrace(request, response, mockNext);
+      await RaceController.showTraitsForRace(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -150,7 +180,7 @@ describe('showTraitsForSubrace', () => {
   });
 });
 
-describe('showProficienciesForSubrace', () => {
+describe('showProficienciesForRace', () => {
   const findDoc = [
     {
       index: 'daggers',
@@ -168,12 +198,12 @@ describe('showProficienciesForSubrace', () => {
       url: '/api/proficiencies/quarterstaffs',
     },
   ];
-  const request = mockRequest({ params: { index: 'high-elf' } });
+  const request = mockRequest({ params: { index: 'elf' } });
 
   it('returns a list of objects', async () => {
     mockingoose(Proficiency).toReturn(findDoc, 'find');
 
-    await SubraceController.showProficienciesForSubrace(request, response, mockNext);
+    await RaceController.showProficienciesForRace(request, response, mockNext);
     expect(response.status).toHaveBeenCalledWith(200);
   });
 
@@ -182,7 +212,7 @@ describe('showProficienciesForSubrace', () => {
       const error = new Error('Something went wrong');
       mockingoose(Proficiency).toReturn(error, 'find');
 
-      await SubraceController.showProficienciesForSubrace(request, response, mockNext);
+      await RaceController.showProficienciesForRace(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
