+++ src/tests/controllers/api/traitController.test.ts
@@ -1,10 +1,11 @@
+import mockingoose from 'mockingoose';
 import { mockNext, mockRequest, mockResponse } from '../../support/requestHelpers.js';
 
-import Condition from '../../../models/condition/index.js';
-import ConditionController from '../../../controllers/api/conditionController.js';
-import mockingoose from 'mockingoose';
+import Trait from '../../../models/trait/index.js';
+import TraitController from '../../../controllers/api/traitController.js';
+import { MockResponse } from '../../support/types.d';
 
-let response;
+let response: MockResponse;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
@@ -13,27 +14,27 @@ beforeEach(() => {
 describe('index', () => {
   const findDoc = [
     {
-      index: 'blinded',
-      name: 'Blinded',
-      url: '/api/conditions/blinded',
+      index: 'artificers-lore',
+      name: "Artificer's Lore",
+      url: '/api/traits/artificers-lore',
     },
     {
-      index: 'charmed',
-      name: 'Charmed',
-      url: '/api/conditions/charmed',
+      index: 'brave',
+      name: 'Brave',
+      url: '/api/traits/brave',
     },
     {
-      index: 'deafened',
-      name: 'Deafened',
-      url: '/api/conditions/deafened',
+      index: 'breath-weapon',
+      name: 'Breath Weapon',
+      url: '/api/traits/breath-weapon',
     },
   ];
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(Condition).toReturn(findDoc, 'find');
+    mockingoose(Trait).toReturn(findDoc, 'find');
 
-    await ConditionController.index(request, response, mockNext);
+    await TraitController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -41,9 +42,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Condition).toReturn(error, 'find');
+      mockingoose(Trait).toReturn(error, 'find');
 
-      await ConditionController.index(request, response, mockNext);
+      await TraitController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -54,18 +55,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'blinded',
-    name: 'Blinded',
-    url: '/api/conditions/blinded',
+    index: 'abyssal',
+    name: 'Abyssal',
+    url: '/api/traits/abyssal',
   };
 
-  const showParams = { index: 'blinded' };
+  const showParams = { index: 'abyssal' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(Condition).toReturn(findOneDoc, 'findOne');
+    mockingoose(Trait).toReturn(findOneDoc, 'findOne');
 
-    await ConditionController.show(request, response, mockNext);
+    await TraitController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -73,11 +74,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(Condition).toReturn(null, 'findOne');
+      mockingoose(Trait).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await ConditionController.show(invalidRequest, response, mockNext);
+      await TraitController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -88,9 +89,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Condition).toReturn(error, 'findOne');
+      mockingoose(Trait).toReturn(error, 'findOne');
 
-      await ConditionController.show(request, response, mockNext);
+      await TraitController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
