+++ src/tests/controllers/simpleController.test.ts
@@ -1,39 +1,42 @@
-import { mockNext, mockRequest, mockResponse } from '../../support/requestHelpers.js';
+import { MockResponse } from '../support/types.d';
+import { mockNext, mockRequest, mockResponse } from '../support/requestHelpers.js';
 
-import WeaponProperty from '../../../models/weaponProperty/index.js';
-import WeaponPropertyController from '../../../controllers/api/weaponPropertyController.js';
+import AbilityScore from '../../models/abilityScore/index.js';
+import SimpleController from '../../controllers/simpleController.js';
 import mockingoose from 'mockingoose';
 
-let response;
+let response: MockResponse;
+let simpleController: SimpleController;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
+  simpleController = new SimpleController(AbilityScore);
 });
 
 describe('index', () => {
   const findDoc = [
     {
-      index: 'ammunition',
-      name: 'Ammunition',
-      url: '/api/weapon-properties/ammunition',
+      index: 'str',
+      name: 'STR',
+      url: '/api/ability-scores/str',
     },
     {
-      index: 'finesse',
-      name: 'Finesse',
-      url: '/api/weapon-properties/finesse',
+      index: 'dex',
+      name: 'DEX',
+      url: '/api/ability-scores/dex',
     },
     {
-      index: 'heavy',
-      name: 'Heavy',
-      url: '/api/weapon-properties/heavy',
+      index: 'con',
+      name: 'CON',
+      url: '/api/ability-scores/con',
     },
   ];
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(WeaponProperty).toReturn(findDoc, 'find');
+    mockingoose(AbilityScore).toReturn(findDoc, 'find');
 
-    await WeaponPropertyController.index(request, response, mockNext);
+    await simpleController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -41,9 +44,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(WeaponProperty).toReturn(error, 'find');
+      mockingoose(AbilityScore).toReturn(error, 'find');
 
-      await WeaponPropertyController.index(request, response, mockNext);
+      await simpleController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -54,18 +57,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'ammunition',
-    name: 'Ammunition',
-    url: '/api/weapon-properties/ammunition',
+    index: 'str',
+    name: 'STR',
+    url: '/api/ability-scores/str',
   };
 
-  const showParams = { index: 'ammunition' };
+  const showParams = { index: 'str' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(WeaponProperty).toReturn(findOneDoc, 'findOne');
+    mockingoose(AbilityScore).toReturn(findOneDoc, 'findOne');
 
-    await WeaponPropertyController.show(request, response, mockNext);
+    await simpleController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -73,11 +76,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(WeaponProperty).toReturn(null, 'findOne');
+      mockingoose(AbilityScore).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await WeaponPropertyController.show(invalidRequest, response, mockNext);
+      await simpleController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -88,9 +91,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(WeaponProperty).toReturn(error, 'findOne');
+      mockingoose(AbilityScore).toReturn(error, 'findOne');
 
-      await WeaponPropertyController.show(request, response, mockNext);
+      await simpleController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
